requirements.txt
fastapi
uvicorn
openai
kokoro
soundfile
python-dotenv
